+++ src/main/java/ca/serum390/godzilla/api/handlers/PlannedProductHandler.java
@@ -26,101 +26,10 @@ public class PlannedProductHandler {
      * {@link ca.serum390.godzilla.data.repositories.PlannedProductsRepository}
      */
     private final PlannedProductsRepository plannedProducts;
-    private final InventoryRepository inventoryRepository;
-    private final OrdersRepository ordersRepository;
 
-    public PlannedProductHandler(PlannedProductsRepository plannedProducts, InventoryRepository inventoryRepository, OrdersRepository ordersRepository) {
+    public PlannedProductHandler(PlannedProductsRepository plannedProducts) {
         this.plannedProducts = plannedProducts;
-        this.inventoryRepository = inventoryRepository;
-        this.ordersRepository = ordersRepository;
     }
-
-    // preproduction  \=month?day?year?
-
-    public Mono<ServerResponse> validateProduction(ServerRequest req) {
-        Order order = req.bodyToMono(Order.class).block();
-        AtomicBoolean isOrderReady = new AtomicBoolean(true);
-        AtomicBoolean isOrderBlocked = new AtomicBoolean(false);
-        Order purchaseOrder = new Order();
-        purchaseOrder.setItems(new HashMap<>());
-        purchaseOrder.setOrderType("purchase");
-        purchaseOrder.setStatus("new");
-        purchaseOrder.setCreatedDate(LocalDate.of(2021, 1, 1));
-        purchaseOrder.setDueDate(LocalDate.of(2021, 1, 1));
-        purchaseOrder.setDeliveryLocation("Montreal");
-        AtomicReference<Integer> productionID = new AtomicReference<>(0);
-        AtomicReference<PlannedProduct> plannedProduct = new AtomicReference<>(new PlannedProduct());
-        int year = 1, month = 1, day = 1;
-        LocalDate productionDate = LocalDate.of(year, month, day);
-
-
-        order.getItems().forEach((itemID, quantity) -> {
-            Item item = inventoryRepository.findById(itemID).block();
-            if (quantity <= item.getQuantity()) {
-                // check inventory for finished item
-                int newQuantity = item.getQuantity() - quantity;
-                System.out.println("Item " + item.getItemName() + " is available");
-                inventoryRepository.update(itemID, newQuantity).block();
-                System.out.println("inventory updated");
-
-            } else {
-                isOrderReady.set(false);
-                plannedProduct.set(new PlannedProduct());
-                plannedProduct.get().setProductionDate(productionDate);
-                plannedProduct.get().setStatus("new");
-                plannedProduct.get().setOrderID(order.getId());
-
-                quantity = quantity - item.getQuantity();
-                inventoryRepository.update(itemID, 0).block();
-                System.out.println("inventory updated with 0");
-                // check bom for that item
-                System.out.println("Item " + item.getItemName() + " is not available");
-                Integer finalQuantity = quantity;
-                item.getBillOfMaterial().forEach((id2, quantity2) -> {
-                            Item item2 = inventoryRepository.findById(id2).block();
-                            int total_quantity = quantity2 * finalQuantity;
-                            if (total_quantity <= item2.getQuantity()) {
-                                // check inventory for finished item
-                                int newQuantity = item2.getQuantity() - total_quantity;
-                                System.out.println("Item " + item2.getItemName() + " is available");
-                                inventoryRepository.update(id2, newQuantity).block();
-                                System.out.println("inventory updated for bom");
-                            } else {
-                                total_quantity = total_quantity - item2.getQuantity();
-                                isOrderBlocked.set(true);
-                                //update inventory
-                                inventoryRepository.update(id2, 0).block();
-                                System.out.println("inventory updated with 0 for bom");
-                                // create purchase order this bom item with the number = total-quantity
-                                purchaseOrder.getItems().put(id2, total_quantity);
-                                System.out.println("purchase order updated");
-                            }
-                        }
-                );
-            }
-        });
-
-        if (isOrderReady.get()) {
-            ordersRepository.update(order.getId(), "Ready")
-                    .subscribe(num -> System.out.println("order status updated"));
-        } else {
-            ordersRepository.update(order.getId(), isOrderBlocked.get() ? "blocked" : "processing")
-                    .subscribe(num -> System.out.println("order status updated"));
-            plannedProducts.save(plannedProduct.get()).subscribe(product ->
-            {
-                System.out.println("production is saved");
-                productionID.set(product.getId());
-                if (isOrderBlocked.get()) {
-                    ordersRepository
-                            .save(purchaseOrder.getCreatedDate(), purchaseOrder.getDueDate(), purchaseOrder.getDeliveryLocation()
-                                    , purchaseOrder.getOrderType(), purchaseOrder.getStatus(), purchaseOrder.getItems(), productionID.get())
-                            .subscribe(Order -> System.out.println("order is saved"));
-                }
-            });
-        }
-        return noContent().build();
-    }
-
 
     /**
      * Insert into the table
