+++ src/main/java/ca/serum390/godzilla/api/routers/ProductionManagerRouter.java
@@ -1,6 +1,7 @@
 package ca.serum390.godzilla.api.routers;
 
 import ca.serum390.godzilla.api.handlers.PlannedProductHandler;
+import ca.serum390.godzilla.api.handlers.ProductionManagerHandler;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.reactive.function.server.RouterFunction;
@@ -8,17 +9,13 @@ import org.springframework.web.reactive.function.server.RouterFunctions;
 import org.springframework.web.reactive.function.server.ServerResponse;
 
 @Configuration
-public class PlannedProductRouter {
+public class ProductionManagerRouter {
     @Bean
-    public RouterFunction<ServerResponse> plannedProductRoute(PlannedProductHandler plannedProductHandler) {
+    public RouterFunction<ServerResponse> productionManagerRoute(ProductionManagerHandler productionManagerHandler) {
         final String ID = "/{id}";
         return RouterFunctions.route()
-                .path("/planning/", builder -> builder
-                        .GET("/", plannedProductHandler::getBy)
-                        .GET(ID, plannedProductHandler::getByID)
-                        .POST("/", plannedProductHandler::create)
-                        .PUT(ID, plannedProductHandler::update)
-                        .DELETE(ID, plannedProductHandler::delete))
+                .path("/production-manager/", builder -> builder
+                        .POST("/validate", productionManagerHandler::validateProduction))
                 .build();
     }
 }
